state_historyを保存できるように
