+++ frontend/packs/application.js
@@ -7,5 +7,4 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-import "init";
-import "components/page/page";
+// import "init";
